refactor(firestore): workProgress.tsを機能別モジュールに分割
773行の単一ファイルをhelpers/crud/progressの3モジュールに分割し、
ステータス遷移・再計算ロジックの重複を共通ヘルパーに集約。
