+++ lib/firestore/workProgress/index.ts
@@ -1,0 +1,11 @@
+// workProgress barrel export
+export { extractTargetAmount, extractUnitFromWeight } from './helpers';
+export { addWorkProgress, updateWorkProgress, updateWorkProgresses, deleteWorkProgress } from './crud';
+export {
+  addCompletedCountToWorkProgress,
+  addProgressToWorkProgress,
+  archiveWorkProgress,
+  unarchiveWorkProgress,
+  updateProgressHistoryEntry,
+  deleteProgressHistoryEntry,
+} from './progress';
